- NEW: Removed the old DataImportingDialog code. - CHG: Some very minor changes to the map importing code to reduce memory usage. - NEW: Updated and changed some of the i18n text used during importing.

+++ trunk/src/flapjack/data/ChromosomeMap.java
@@ -134,12 +134,12 @@ public class ChromosomeMap extends XMLRoot implements Iterable<Marker>
 	public static class Wrapper
 	{
 		public ChromosomeMap map;
-		public int index;
+		public short index;
 
 		public Wrapper(ChromosomeMap map, int index)
 		{
 			this.map = map;
-			this.index = index;
+			this.index = (short) index;
 		}
 	}
 }
